Incorporate UnknownPlatformTests and UpcomingFeatureTests into IfConfigExprTests
